Videos re-encoded
I have tried to play the first video from
https://hollo.me/php/experimental-async-php-volume-2.html and have
noticed a rather abnormal behavior with HTML5 video player.

It would consume 100% CPU (all 4 cores) on Chromium under Debian testing
64-bit and nearly 80% CPU (again, all 4 cores) on Firefox.

The original video's resolution was '2824x1708' and I reduced it down to
'1280x720' for both videos and optimized it for web.

Both have been tested and perform up to 5% the most.